refactor: Image, Axios FSD 구조 변경
