+++ src/entities/comment/api/get-comments.ts
@@ -1,4 +1,4 @@
-import { customAxios } from '@/src/shared/axios'
+import { customAxios } from '@/src/shared/api'
 import { CommentType } from '../model'
 import { useQuery } from '@tanstack/react-query'
 import { COMMENT_QUERY_KEY } from './queryKey'
