+++ src/entities/comment/api/update-comment.ts
@@ -1,4 +1,4 @@
-import { customAxios } from '@/src/shared/axios'
+import { customAxios } from '@/src/shared/api'
 import { CommentType } from '../model'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { COMMENT_URL } from './endpoint'
