+++ src/entities/notification/api/get-notifications.ts
@@ -1,4 +1,4 @@
-import { customAxios } from '@/src/shared/axios'
+import { customAxios } from '@/src/shared/api'
 import {
   NotificationType,
   NOTIFICATION_URL,
