+++ src/entities/user/api/delete-my-region-preferences.ts
@@ -1,4 +1,4 @@
-import { customAxios } from '@/src/shared/axios'
+import { customAxios } from '@/src/shared/api'
 import { USER_API } from './endpoint'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { USER_QUERY_KEY } from './queryKey'
