+++ src/features/place/form-review.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 import React, { useState, useRef, useEffect } from 'react'
 import { FieldErrors, UseFormRegister, UseFormSetValue } from 'react-hook-form'
 import { ReviewPayloadType } from '@/src/entities/place'
-import { postImage } from '@/src/shared/entities/api'
+import { postImage } from '@/src/shared/api'
 import { message } from 'antd'
 import Spacer from '@/src/shared/ui/Spacer'
