+++ src/shared/api/image/index.ts
@@ -1,0 +1,2 @@
+export * from './get-image-upload-url'
+export * from './post-image'
