+++ src/shared/api/index.ts
@@ -1,0 +1,2 @@
+export * from './axios'
+export * from './image'
